Run
調整難度

+++ Assets/william/script/SO/AngerData.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New AngerData", menuName = "Item/Anger Data", order = 1)]
 public class AngerData : GameData
 {
-    public float anger;
-    public float anger_max;
+    public int anger;
+    public int anger_max;
     public bool IsOver()
     {
         return anger <= 0;
